+++ examples/eth-pm-wallet-encryption/src/waku.ts
@@ -11,7 +11,7 @@ export const PrivateMessageContentTopic =
   '/eth-pm-wallet/1/private-message/proto';
 
 export async function initWaku(): Promise<Waku> {
-  const waku = await Waku.create({ bootstrap: true });
+  const waku = await Waku.create({ bootstrap: { default: true } });
 
   // Wait to be connected to at least one peer
   await new Promise((resolve, reject) => {