+++ src/index.ts
@@ -1,4 +1,7 @@
-export { getBootstrapNodes } from './lib/discovery';
+export { getNodesFromHostedJson } from './lib/discovery';
+export * as discovery from './lib/discovery';
+
+export * as enr from './lib/enr';
 
 export * as utils from './lib/utils';